requirements.txt
requests
beautifulsoup4
pygame
python-bps-continued